% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_events.R
\name{af_add_event_info}
\alias{af_add_event_info}
\title{Add additional event information (like Event Types) to Survey Data}
\usage{
af_add_event_info(df, wave_var, wave_order, event_table, community_var)
}
\arguments{
\item{df}{A dataframe containing all survey data}

\item{wave_var}{Character string specifying the name of the variable that identifies survey waves}

\item{wave_order}{Vector mapping wave numbers to wave names (e.g., c("First", "Second", "Third", ...))}

\item{event_table}{Dataframe containing event additional information}

\item{community_var}{Character string specifying the community variable name to match with event_table columns}
}
\value{
Modified dataframe with additional event information columns
}
\description{
This function adds event_result and event_type columns to the entire dataframe based on
event_table and community variable mapping. The first wave gets "No-Event" and
all other waves get values based on the corresponding event in event_table.
}
