% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_weights.R
\name{af_analyze_extremism_weighted}
\alias{af_analyze_extremism_weighted}
\title{Analyze extremism differences with equal group weighting}
\usage{
af_analyze_extremism_weighted(
  data,
  outcome_var,
  group_var,
  wave_var,
  weight_var = "equal_weight"
)
}
\arguments{
\item{data}{Weighted data frame}

\item{outcome_var}{Your extremism outcome variable name}

\item{group_var}{Political group variable}

\item{wave_var}{Wave variable}

\item{weight_var}{Weight variable name}
}
\description{
Analyze extremism differences with equal group weighting
}
