% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_gauge_indices.R
\name{af_calculate_concentration}
\alias{af_calculate_concentration}
\title{Calculate extremism concentration ratio}
\usage{
af_calculate_concentration(df, comm_var, numerator_metric, denominator_metric)
}
\arguments{
\item{df}{A dataframe containing the extremism data}

\item{comm_var}{Character string representing the community variable column name}

\item{numerator_metric}{The column name of the numerator metric (e.g., er3_c)}

\item{denominator_metric}{The column name of the denominator metric (e.g., er1_c)}
}
\value{
A dataframe with the calculated concentration ratios
}
\description{
Calculate extremism concentration ratio
}
