% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_gauge_indices.R
\name{af_calculate_ratio}
\alias{af_calculate_ratio}
\title{Calculate ratio of community metric to population metric}
\usage{
af_calculate_ratio(df, comm_var, community_metric, population_metric)
}
\arguments{
\item{df}{A dataframe containing the extremism data}

\item{comm_var}{Character string representing the community variable column name}

\item{community_metric}{The column name of the community metric}

\item{population_metric}{The column name of the population metric}
}
\value{
A dataframe with the calculated ratios
}
\description{
Calculate ratio of community metric to population metric
}
