% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_weights.R
\name{af_create_equal_weights}
\alias{af_create_equal_weights}
\title{Create weights for equal group sizes within each wave}
\usage{
af_create_equal_weights(data, wave_var, group_var)
}
\arguments{
\item{data}{Survey data frame}

\item{wave_var}{Name of wave variable}

\item{group_var}{Name of grouping variable}
}
\value{
Data frame with added weight variable
}
\description{
Create weights for equal group sizes within each wave
}
