% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_least_liked.R
\name{af_create_leastliked_table}
\alias{af_create_leastliked_table}
\title{Create percentage table of least liked groups by wave}
\usage{
af_create_leastliked_table(data, political_orientation = NULL, title = "")
}
\arguments{
\item{data}{A dataframe containing survey data}

\item{political_orientation}{Optional character string to filter by political orientation}

\item{title}{Character string for table title}
}
\value{
A gt table object
}
\description{
Create percentage table of least liked groups by wave
}
