% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_weights.R
\name{af_create_pop_weights}
\alias{af_create_pop_weights}
\title{Create population-representative weights}
\usage{
af_create_pop_weights(data, wave_var, group_var, pop_props)
}
\arguments{
\item{data}{Survey data frame}

\item{wave_var}{Name of wave variable}

\item{group_var}{Name of grouping variable}

\item{pop_props}{Named vector of population proportions (must sum to 1)}
}
\value{
Data frame with added weight variable
}
\description{
Create population-representative weights
}
