% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_extremists.R
\name{af_extremists}
\alias{af_extremists}
\title{Plot Characteristics of Extremist Groups in Populations or Communities}
\usage{
af_extremists(
  df,
  extremism_level,
  characteristics,
  pop_var1 = NULL,
  pop_levels1 = NULL,
  pop_var2 = NULL,
  pop_levels2 = NULL,
  pop_var3 = NULL,
  pop_levels3 = NULL,
  comm_var1 = NULL,
  comm_levels1 = NULL,
  comm_var2 = NULL,
  comm_levels2 = NULL,
  type = "by_characteristic",
  bar_type = "Vertical",
  ncol = 2
)
}
\arguments{
\item{df}{A data frame containing the survey dataset with extremist indicators (output from af_gauge_indices)}

\item{extremism_level}{Integer indicating which extremist definition to use (1, 2, or 3)}

\item{characteristics}{Character vector of variable names to plot as characteristics}

\item{pop_var1}{Optional first population variable name}

\item{pop_levels1}{Optional list of levels for pop_var1, if NULL all levels are used}

\item{pop_var2}{Optional second population variable name}

\item{pop_levels2}{Optional list of levels for pop_var2, if NULL all levels are used}

\item{pop_var3}{Optional third population variable name}

\item{pop_levels3}{Optional list of levels for pop_var3, if NULL all levels are used}

\item{comm_var1}{Optional first community variable name}

\item{comm_levels1}{Optional list of levels for comm_var1, if NULL all levels are used}

\item{comm_var2}{Optional second community variable name}

\item{comm_levels2}{Optional list of levels for comm_var2, if NULL all levels are used}

\item{type}{Character string indicating plot organization, either "by_characteristic" or "by_group"}

\item{bar_type}{Character string specifying bar orientation, either "Vertical" or "Horizontal"}

\item{ncol}{Integer specifying the number of columns in the combined plot}
}
\value{
A ggplot object containing all created plots combined
}
\description{
Plot Characteristics of Extremist Groups in Populations or Communities
}
