% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_manova.R
\name{af_format_p_value}
\alias{af_format_p_value}
\title{Format P-values for Display}
\usage{
af_format_p_value(p_value)
}
\arguments{
\item{p_value}{Numeric p-value}
}
\value{
Character string of formatted p-value
}
\description{
Formats p-values for display in tables, handling very small values.
}
