% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_did.R
\name{af_format_rates_gt}
\alias{af_format_rates_gt}
\title{Format Extremism Rates as GT Table}
\usage{
af_format_rates_gt(rates_data)
}
\arguments{
\item{rates_data}{Data frame from af_calculate_extremism_rates function}
}
\value{
gt table object with formatted extremism rates
}
\description{
Creates a formatted gt table from extremism rates data
}
