% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_gauge_indices.R
\name{af_gauge_indices}
\alias{af_gauge_indices}
\title{Calculate Political Extremism Gauge Indices}
\usage{
af_gauge_indices(
  df,
  pop_var1 = NULL,
  pop_levels1 = NULL,
  pop_var2 = NULL,
  pop_levels2 = NULL,
  pop_var3 = NULL,
  pop_levels3 = NULL,
  comm_var1 = NULL,
  comm_levels1 = NULL,
  comm_var2 = NULL,
  comm_levels2 = NULL,
  dimensions = c("pe_ideology", "pe_violence", "pe_intolerance"),
  threshold_type = "MAD",
  k_factor = 1.5
)
}
\arguments{
\item{df}{A data frame containing the survey dataset}

\item{pop_var1}{Optional first population variable name}

\item{pop_levels1}{Optional list of levels for pop_var1, if NULL all levels are used}

\item{pop_var2}{Optional second population variable name}

\item{pop_levels2}{Optional list of levels for pop_var2, if NULL all levels are used}

\item{pop_var3}{Optional third population variable name}

\item{pop_levels3}{Optional list of levels for pop_var3, if NULL all levels are used}

\item{comm_var1}{Optional first community variable name}

\item{comm_levels1}{Optional list of levels for comm_var1, if NULL all levels are used}

\item{comm_var2}{Optional second community variable name}

\item{comm_levels2}{Optional list of levels for comm_var2, if NULL all levels are used}

\item{dimensions}{Character vector of exactly 3 variable names corresponding to cognitive, behavioral, and social dimensions}

\item{threshold_type}{Character string specifying the threshold calculation method. Must be either "MAD", "Qn", "Sn" or "Tau". Default is "MAD".}

\item{k_factor}{Numeric value specifying the multiplier for MAD method. Must be positive. Default is 2.}
}
\value{
A list containing two elements:
\item{df}{The original data frame with added columns for the calculated indices}
\item{indices_table}{A data frame with one row per population/community combination and their calculated indices}
}
\description{
Calculate Political Extremism Gauge Indices
}
