% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_anova.R
\name{af_generate_results_text}
\alias{af_generate_results_text}
\title{Generate Results Text Summary}
\usage{
af_generate_results_text(
  anova_results,
  descriptive_stats,
  extremism_var,
  group_var = NULL,
  control_vars = NULL
)
}
\arguments{
\item{anova_results}{List of ANOVA results for each wave pair}

\item{descriptive_stats}{List of descriptive statistics}

\item{extremism_var}{Name of the extremism variable}

\item{group_var}{Name of the group variable}

\item{control_vars}{Names of control variables (if any)}
}
\value{
Character string with formatted results
}
\description{
Internal function to generate formatted results text similar to academic papers
}
