% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_gauge_indices.R
\name{af_get_available_communities}
\alias{af_get_available_communities}
\title{Get Available Communities}
\usage{
af_get_available_communities(data, pop_var, comm_var, selected_pops = NULL)
}
\arguments{
\item{data}{Data frame containing the survey data}

\item{pop_var}{Character string representing the population variable column name}

\item{comm_var}{Character string representing the community variable column name}

\item{selected_pops}{Vector of selected population values (optional)}
}
\value{
Vector of available community values
}
\description{
Retrieves available community values based on selected populations
}
