% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_indices_table.R
\name{af_indices_table}
\alias{af_indices_table}
\title{Format and Print Political Extremism Gauge Indices Table}
\usage{
af_indices_table(indices_table, print_type = NULL)
}
\arguments{
\item{indices_table}{A data frame containing the indices table created by af_gauge_indices}

\item{print_type}{What to print: NULL for all, "Population" for population only,
"Community" for community only, or a character vector of specific indices}
}
\value{
A gt table object
}
\description{
Format and Print Political Extremism Gauge Indices Table
}
