% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_plot_gauge.R
\name{af_plot_gauge}
\alias{af_plot_gauge}
\title{Create 3D Political Extremism Gauge Plot}
\usage{
af_plot_gauge(
  indices_table,
  indices_group = "NP",
  rank_index = "ER1",
  pop_var1 = NULL,
  pop_levels1 = NULL,
  pop_var2 = NULL,
  pop_levels2 = NULL,
  pop_var3 = NULL,
  pop_levels3 = NULL,
  comm_var1 = NULL,
  comm_levels1 = NULL,
  comm_var2 = NULL,
  comm_levels2 = NULL
)
}
\arguments{
\item{indices_table}{Data frame containing the indices table from af_gauge_indices}

\item{indices_group}{Type of indices to use: "NP" (Normative Points), "EP" (Extremism Points), or "EL" (Extremism Levels)}

\item{rank_index}{Extremism Rank index to use for coloring: "ER1", "ER2", or "ER3"}

\item{pop_var1}{Optional first population variable name}

\item{pop_levels1}{Optional list of levels for pop_var1}

\item{pop_var2}{Optional second population variable name}

\item{pop_levels2}{Optional list of levels for pop_var2}

\item{pop_var3}{Optional third population variable name}

\item{pop_levels3}{Optional list of levels for pop_var3}

\item{comm_var1}{Optional first community variable name}

\item{comm_levels1}{Optional list of levels for comm_var1}

\item{comm_var2}{Optional second community variable name}

\item{comm_levels2}{Optional list of levels for comm_var2}
}
\value{
A plotly figure object representing a 3D scatter plot
}
\description{
Create 3D Political Extremism Gauge Plot
}
