% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_weights.R
\name{af_plot_group_trajectories}
\alias{af_plot_group_trajectories}
\title{Plot group trajectories across waves}
\usage{
af_plot_group_trajectories(
  results,
  group_var = "pe_left_center_right",
  wave_var = "Wave",
  outcome_name = "Extremism Score"
)
}
\arguments{
\item{results}{Output from af_analyze_extremism_weighted}

\item{group_var}{Name of grouping variable (for proper legend labeling)}

\item{wave_var}{Name of wave variable (for proper x-axis labeling)}

\item{outcome_name}{Readable name for outcome variable (for y-axis label)}
}
\description{
Plot group trajectories across waves
}
