% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_plot_indices.R
\name{af_plot_indices}
\alias{af_plot_indices}
\title{Plot Political Extremism Gauge Indices}
\usage{
af_plot_indices(
  indices_table,
  selected_indices = NULL,
  pop_var1 = NULL,
  pop_levels1 = NULL,
  pop_var2 = NULL,
  pop_levels2 = NULL,
  pop_var3 = NULL,
  pop_levels3 = NULL,
  comm_var1 = NULL,
  comm_levels1 = NULL,
  comm_var2 = NULL,
  comm_levels2 = NULL,
  type = "by_index",
  bar_type = "Vertical",
  ncol = 2
)
}
\arguments{
\item{indices_table}{A data frame containing the indices table from af_gauge_indices}

\item{selected_indices}{Character vector of indices to plot, if NULL all indices will be selected}

\item{pop_var1}{Optional first population variable name}

\item{pop_levels1}{Optional list of levels for pop_var1}

\item{pop_var2}{Optional second population variable name}

\item{pop_levels2}{Optional list of levels for pop_var2}

\item{pop_var3}{Optional third population variable name}

\item{pop_levels3}{Optional list of levels for pop_var3}

\item{comm_var1}{Optional first community variable name}

\item{comm_levels1}{Optional list of levels for comm_var1}

\item{comm_var2}{Optional second community variable name}

\item{comm_levels2}{Optional list of levels for comm_var2}

\item{type}{Type of plot to generate: "by_index" (default) or "by_group"}

\item{bar_type}{Orientation of the bars: "Vertical" (default) or "Horizontal"}

\item{ncol}{Number of plots per row in the combined plot}
}
\value{
A ggplot object with the combined plots
}
\description{
Plot Political Extremism Gauge Indices
}
