% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_gauge_indices.R
\name{af_update_index_names}
\alias{af_update_index_names}
\title{Update Index Names Based on Community Selection}
\usage{
af_update_index_names(index_names, has_community)
}
\arguments{
\item{index_names}{List of index names}

\item{has_community}{Logical indicating whether a community variable is selected}
}
\value{
Updated list of index names with appropriate suffixes
}
\description{
Updates index names with appropriate suffix (_p or _c) based on community selection
}
